requirements-local.txt
# ============================================================================
# IPO Compliance System - Local Requirements (No GCP)
# ============================================================================

# --- Web ---
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# --- DB ---
neo4j==5.15.0

# --- Core ---
python-dotenv==1.0.0
numpy>=1.26.2
pandas==2.3.0
PyPDF2==3.0.1
tqdm==4.66.5
pydantic==2.9.2

# --- LangChain Stack ---
langgraph==0.2.45
langchain==0.3.7
langchain-community==0.3.7

# --- AI/ML (Local alternatives) ---

# Google Gemini
google-generativeai>=0.8.0 
google-genai>=1.50.0

# Option 1: Use OpenAI API
openai>=1.0.0

# Option 2: Use Anthropic Claude API
anthropic>=0.18.0

# Option 3: Use Ollama (fully local)
# ollama-python>=0.1.0

# --- File Storage (Local) ---
# No GCS needed - use local filesystem


# --- Sentence Transformers (for embeddings if needed) ---
sentence-transformers>=2.2.0

# --- Other utilities ---
requests>=2.31.0

camelot-py[cv]  # For bordered tables
tabula-py       # For stream tables